move root app component #20
