+++ packages/frontend/src/app.tsx
@@ -1,13 +1,14 @@
 import { getConfiguration } from 'elite-configuration';
-import { FeatureFlagsProvider } from 'elite-feature-flags';
-import { Configuration } from 'elite-types';
+import { AppPath, Configuration } from 'elite-types';
 import * as React from 'react';
 import { hot } from 'react-hot-loader';
 import { Redirect, Route, Switch } from 'react-router';
 import { Router } from 'react-router-dom';
-import history from '../util/history';
-import { ErrorBoundary } from './general/ErrorBoundary';
-import { AppPath, APP_ROUTES } from '../util/routes';
+import history from './util/history';
+import { APP_ROUTES } from './util/routes';
+import { ErrorBoundaryComponent, FeatureFlagsProvider } from 'elite-components';
+import { Divider } from '@material-ui/core';
+import { LinkDirectory } from './util/linkDirectory';
 
 // Global bootstrap: install subsystems and load configuration
 const configuration: Configuration = getConfiguration();
@@ -15,15 +16,17 @@ const configuration: Configuration = getConfiguration();
 export const AppComponent = () => (
   <FeatureFlagsProvider value={configuration.featureMap}>
     <Router history={history}>
-      <Switch>
-        {APP_ROUTES.map((routeProps, index) => (
-          <ErrorBoundary>
+      <ErrorBoundaryComponent>
+        <Switch>
+          {APP_ROUTES.map((routeProps, index) => (
             <Route key={index} {...routeProps} />
-          </ErrorBoundary>
-        ))}
-        {/* Error 404 Fallback */}
-        <Redirect to={AppPath.HOME} />
-      </Switch>
+          ))}
+          {/* Error 404 Fallback */}
+          <Redirect to={AppPath.HOME} />
+        </Switch>
+        <Divider />
+        <LinkDirectory />
+      </ErrorBoundaryComponent>
     </Router>
   </FeatureFlagsProvider>
 );
