+++ packages/frontend/src/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { App } from './components/App';
+import { App } from './app';
 
 ReactDOM.render(<App />, document.getElementById('root'));
